refactor(oidc): share token credential and scope parsing helpers
Extract shared Basic-auth client credential parsing and scope normalization helpers to remove duplicated token exchange logic and satisfy Sonar new-code duplication limits.
