+++ src/app/api/signer/device/exchange/route.ts
@@ -11,22 +11,7 @@ import {
   resolveSubjectAccessToken,
   SubjectAccessTokenResolveError,
 } from "@/lib/oidc/resolve-subject-access-token";
-
-function scopeStringFromPayload(payload: Record<string, unknown>): string {
-  const scopeFromScope =
-    typeof payload.scope === "string" ? payload.scope.trim() : "";
-  if (scopeFromScope) {
-    return scopeFromScope.replace(/\s+/g, " ").trim();
-  }
-  const scpRaw = payload.scp;
-  if (Array.isArray(scpRaw)) {
-    return scpRaw.filter((v): v is string => typeof v === "string").join(" ");
-  }
-  if (typeof scpRaw === "string") {
-    return scpRaw.trim();
-  }
-  return "sign:job";
-}
+import { scopeStringFromPayload } from "@/lib/oidc/scope-string";
 
 async function mintFromDeviceToken(
   deviceToken: string,
