+++ src/lib/oidc/signer-jwt-token-exchange.ts
@@ -18,6 +18,7 @@ import {
   subjectAccessTokenResolveErrorToTokenExchange,
   SubjectAccessTokenResolveError,
 } from "@/lib/oidc/resolve-subject-access-token";
+import { scopeStringFromPayload } from "@/lib/oidc/scope-string";
 import { TokenExchangeError } from "@/lib/oidc/token-exchange";
 
 export const SUBJECT_ACCESS_TOKEN_TYPE =
@@ -94,22 +95,6 @@ export function isSignerJwtTokenExchangeRequest(params: {
     params.subjectTokenType.trim() === SUBJECT_ACCESS_TOKEN_TYPE &&
     targetsSignerJwtAudience(params)
   );
-}
-
-function scopeStringFromPayload(payload: Record<string, unknown>): string {
-  const scopeFromScope =
-    typeof payload.scope === "string" ? payload.scope.trim() : "";
-  if (scopeFromScope) {
-    return scopeFromScope.replace(/\s+/g, " ").trim();
-  }
-  const scpRaw = payload.scp;
-  if (Array.isArray(scpRaw)) {
-    return scpRaw.filter((v): v is string => typeof v === "string").join(" ");
-  }
-  if (typeof scpRaw === "string") {
-    return scpRaw.trim();
-  }
-  return "sign:job";
 }
 
 export async function handleSignerJwtTokenExchange(
